blog_agent/blog/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
distro==1.9.0
Django==5.0.4
django-admin==2.0.2
django-excel-response2==3.0.6
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.15.1
docstring_parser==0.16
excel-base==1.0.4
exceptiongroup==1.2.1
faiss-cpu==1.8.0
frozenlist==1.4.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isoweek==1.3.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.49
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.23.2
orjson==3.10.1
packaging==23.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rsa==4.9
screen==1.0.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
tenacity==8.2.3
tiktoken==0.6.0
TimeConvert==3.0.13
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
xlwt==1.3.0
yarl==1.9.4
